refactor: move NeuronTrainer from base.py to transformers.py
