+++ optimum/neuron/trainers/__init__.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .base import NeuronTrainer
+from .transformers import NeuronTrainer
 from .orpo_trainer import NeuronORPOTrainer
 from .sft_trainer import NeuronSFTTrainer
 from .training_args import NeuronTrainingArguments
